Correction alignement variable val_float
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@2636 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Include/Reseaux_satellite.h
@@ -38,7 +38,7 @@
 	gchar from[12];
     union { gchar string[48];
 		    struct { gint num;
-				     gint val_float;
+				     gfloat val_float;
 				   };
           };
   };
